- maximum cache size for frames can now controlled per movie with movienode attrib "maxcachesize" - default cache size is now 32 frames - correct audio flag evaluation
git-svn-id: https://svn.intern.artcom.de/pro60/trunk/src@18223 f3027e6b-e3f5-0310-a8e1-edde99107618

+++ Y60/components/FFMpegDecoder1/FFMpegDecoder1.cpp
@@ -380,7 +380,7 @@ namespace y60 {
                 if (frameFinished == 0) {
                     continue;
                 }
-                _myLastVideoTimestamp = myPacket.dts;
+                _myLastVideoTimestamp = double(myPacket.dts);
 
                 // suitable framestamp?
                 if (_myLastVideoTimestamp >= theTimestamp) {
@@ -394,7 +394,7 @@ namespace y60 {
         convertFrame(_myFrame, theTargetRaster);
         // store actual timestamp
 
-        theTimestamp = myPacket.dts;
+        theTimestamp = double(myPacket.dts);
         AC_DEBUG << "Packet dts: " << myPacket.dts << ", pts: " << myPacket.pts;
 
         _myLastVStreamIndex = myPacket.stream_index;
